+++ PostHogTests/PostHogAutocaptureEventTrackerSpec.swift
@@ -19,7 +19,7 @@
                     let view = UIView()
                     let eventData = view.eventData!
 
-                    expect(eventData.targetClass).to(equal("UIView"))
+                    expect(eventData.viewHierarchy.first?.targetClass).to(equal("UIView"))
                     expect(eventData.viewHierarchy.count).to(equal(1))
                 }
 
@@ -29,7 +29,7 @@
                     superview.addSubview(button)
                     let eventData = button.eventData!
 
-                    expect(eventData.targetClass).to(equal("UIButton"))
+                    expect(eventData.viewHierarchy.first?.targetClass).to(equal("UIButton"))
                     expect(eventData.viewHierarchy.count).to(equal(2))
                     expect(eventData.screenName).to(beNil())
                 }